variance_inflation_factor: resolve misleading result for non-constant exog
Variance inflation factor (VIF) is used for testing colinearly of multiple features in a linear model, while the linear model may or may not have a constant term.

When the linear model does not have a constant term, users generally do not add constant to `exog`, which provides a misleading R^2 value, and this R^2 value has nothing to do which the colinearly of components in `exog`. In `OLS`, the parameter `hasconstant` resolves this and provide an alternative way to calculate R^2 in that case. Explicitly speaking, if we defined the R^2 in the form of R^2=1-SSE/SST, when there is a constant term, SST should be the sum of (y - mean)^2; when there are no constant term, SST should be the sum of y^2. The R^2 obtained by y^2 is what VIF actually wants. Therefore, it is necessary to provide the `hasconstant` parameter so users can obtain meaningful VIF result at that case.

When the linear model has a constant term, users are responsible for `add_constant`-ing to the `exog`, and everything works well. Explicityly listing out the `hasconstant` reminds the users of their responsibility.

Cf
https://github.com/statsmodels/statsmodels/issues/27
https://github.com/statsmodels/statsmodels/issues/423
https://github.com/statsmodels/statsmodels/pull/499